+++ src/components/data-processing/MitochondrialContent/MitochondrialContent.jsx
@@ -1,58 +1,22 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import _ from 'lodash';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import {
-  Collapse, Row, Col, Space, Skeleton, Divider,
-} from 'antd';
 
 import { generateDataProcessingPlotUuid } from 'utils/generateCustomPlotUuid';
-import {
-  updatePlotConfig,
-  loadPlotConfig,
-  savePlotConfig,
-} from 'redux/actions/componentConfig';
 
+import PlotLayout from 'components/data-processing/PlotLayout';
 import MitochondrialFractionHistogram from '../../plots/MitochondrialFractionHistogram';
 import MitochondrialFractionScatterplot from '../../plots/MitochondrialFractionScatterplot';
 
-import PlotStyling from '../../plots/styling/PlotStyling';
-import MiniPlot from '../../plots/MiniPlot';
-import CalculationConfigContainer from '../CalculationConfigContainer';
 import CalculationConfig from './CalculationConfig';
-import FilterResultTable from '../FilterResultTable';
-
-const { Panel } = Collapse;
 
 const filterName = 'mitochondrialContent';
-
-const allowedPlotActions = {
-  export: true,
-  compiled: false,
-  source: false,
-  editor: false,
-};
 
 const MitochondrialContent = (props) => {
   const {
     experimentId, sampleId, sampleIds, onConfigChange, stepDisabled, stepHadErrors,
   } = props;
 
-  const dispatch = useDispatch();
-
   const [selectedPlot, setSelectedPlot] = useState('histogram');
-  const [plot, setPlot] = useState(null);
-  const filterTableUuid = generateDataProcessingPlotUuid(sampleId, filterName, 2);
-  const filterTableData = useSelector((state) => state.componentConfig[filterTableUuid]?.plotData);
-
-  const debounceSave = useCallback(
-    _.debounce((plotUuid) => dispatch(savePlotConfig(experimentId, plotUuid)), 2000), [],
-  );
-
-  const updatePlotWithChanges = (obj) => {
-    dispatch(updatePlotConfig(plots[selectedPlot].plotUuid, obj));
-    debounceSave(plots[selectedPlot].plotUuid);
-  };
 
   const plots = {
     histogram: {
@@ -83,42 +47,6 @@ const MitochondrialContent = (props) => {
     },
   };
 
-  const selectedConfig = useSelector(
-    (state) => state.componentConfig[plots[selectedPlot].plotUuid]?.config,
-  );
-
-  const expConfig = useSelector(
-    (state) => state.experimentSettings.processing[filterName][sampleId].filterSettings,
-  );
-
-  const selectedPlotData = useSelector(
-    (state) => state.componentConfig[plots[selectedPlot].plotUuid]?.plotData,
-  );
-
-  useEffect(() => {
-    Object.values(plots).forEach((obj) => {
-      if (!selectedConfig) {
-        dispatch(loadPlotConfig(experimentId, obj.plotUuid, obj.plotType));
-      }
-    });
-  }, []);
-
-  useEffect(() => {
-    if (!filterTableData) dispatch(loadPlotConfig(experimentId, filterTableUuid, 'filterTable'));
-  }, []);
-
-  useEffect(() => {
-    if (selectedConfig && selectedPlotData && expConfig) {
-      let newConfig = _.clone(selectedConfig);
-
-      const expConfigSettings = expConfig.methodSettings[expConfig.method];
-
-      newConfig = _.merge(newConfig, expConfigSettings);
-
-      setPlot(plots[selectedPlot].plot(newConfig, selectedPlotData, allowedPlotActions));
-    }
-  }, [expConfig, selectedConfig, selectedPlotData]);
-
   const plotStylingControlsConfig = [
     {
       panelTitle: 'Legend',
@@ -142,90 +70,23 @@ const MitochondrialContent = (props) => {
     },
   ];
 
-  const renderPlot = () => {
-    // Spinner for main window
-    if (!selectedConfig || !selectedPlotData || stepHadErrors) {
-      return (
-        <center>
-          <Skeleton.Image style={{ width: 400, height: 400 }} />
-        </center>
-      );
-    }
-    if (plot) {
-      return plot;
-    }
-  };
+  const renderCalculationConfig = () => <CalculationConfig />;
 
   return (
-    <>
-      <Row gutter={16}>
-        <Col span={18}>
-          <Row>
-            <Col flex='auto'>
-              {renderPlot()}
-            </Col>
-
-            <Col flex='1 0px'>
-              <Space direction='vertical'>
-                {Object.entries(plots).map(([key, plotObj]) => (
-                  <button
-                    type='button'
-                    key={key}
-                    onClick={() => setSelectedPlot(key)}
-                    style={{
-                      margin: 0,
-                      backgroundColor: 'transparent',
-                      align: 'center',
-                      padding: '8px',
-                      border: '1px solid #000',
-                      cursor: 'pointer',
-                    }}
-                  >
-                    <MiniPlot
-                      experimentId={experimentId}
-                      plotUuid={plotObj.plotUuid}
-                      plotFn={plotObj.plot}
-                      actions={false}
-                    />
-                  </button>
-
-                ))}
-              </Space>
-            </Col>
-          </Row>
-          <Divider />
-          <Row style={{ marginTop: '0.5em' }}>
-            {filterTableData
-              ? <FilterResultTable tableData={filterTableData} />
-              : <Skeleton />}
-          </Row>
-        </Col>
-        <Col flex='1 0px'>
-          <Collapse defaultActiveKey='settings'>
-            <Panel header='Filtering Settings' key='settings'>
-              <CalculationConfigContainer
-                filterUuid={filterName}
-                sampleId={sampleId}
-                sampleIds={sampleIds}
-                onConfigChange={onConfigChange}
-                plotType={selectedPlot}
-                stepDisabled={stepDisabled}
-              >
-                <CalculationConfig />
-              </CalculationConfigContainer>
-            </Panel>
-            <Panel header='Plot styling' key='styling'>
-              <div style={{ height: 8 }} />
-              <PlotStyling
-                formConfig={plotStylingControlsConfig}
-                config={selectedConfig}
-                onUpdate={updatePlotWithChanges}
-              />
-            </Panel>
-          </Collapse>
-        </Col>
-      </Row>
-    </>
+    <PlotLayout
+      experimentId={experimentId}
+      plots={plots}
+      setSelectedPlot={setSelectedPlot}
+      selectedPlot={selectedPlot}
+      filterName={filterName}
+      sampleId={sampleId}
+      sampleIds={sampleIds}
+      onConfigChange={onConfigChange}
+      stepDisabled={stepDisabled}
+      plotStylingControlsConfig={plotStylingControlsConfig}
+      renderCalculationConfig={renderCalculationConfig}
+      stepHadErrors={stepHadErrors}
+    />
   );
 };
